% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtmcorr.R
\name{dtmcorr}
\alias{dtmcorr}
\title{Pearson Correlation for Sparse Matrices}
\usage{
dtmcorr(x)
}
\arguments{
\item{x}{A matrix, potentially a sparse matrix such as a "dgCMatrix" object}
}
\value{
a correlation matrix
}
\description{
Pearson Correlation for Sparse Matrices.
More memory and time-efficient than \code{cor(as.matrix(x))}.
}
