% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtmtfidf.R
\name{dtmtfidf}
\alias{dtmtfidf}
\title{Term Frequency - Inverse Document Frequency calculation}
\usage{
dtmtfidf(dtm)
}
\arguments{
\item{dtm}{an object class "dgCMatrix"}
}
\value{
a vector with tfidf values, one for each term in the \code{dtm} matrix
}
\description{
Term Frequency - Inverse Document Frequency calculation.
Averaged by each term.
}
