% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHBiplot.R
\name{GHBiplot}
\alias{GHBiplot}
\title{GHBiplot}
\usage{
GHBiplot (X, Transform.Data = 'scale')
}
\arguments{
\item{X}{array_like; \cr
A data frame which provides the data to be analyzed. All the variables must be numeric.}

\item{Transform.Data}{character; \cr
A value indicating whether the columns of X (variables) should be centered or scaled. The options are: "center" if center is TRUE, centering is done by subtracting the column means (omitting NA) of x from their corresponding columns, and if center is FALSE, centering is not done. "scale" the value of scale determines how column scaling is performed (after centering). If scale is a numeric-alike vector with length equal to the number of columns of x, then each column of x is divided by the corresponding value from scale. If scale is TRUE then scaling is done by dividing the (centered) columns of x by their standard deviations if center is TRUE, and the root mean square otherwise. If scale is FALSE, no scaling is done. To scale by standard deviations without centering, use scale(x,center=FALSE,scale=apply(x,2,sd,na.rm=TRUE)),"center_scale" center=TRUE and scale=TRUE,"none" neither center nor scale is done. The default value is "scale".}
}
\value{
\code{GHBiplot} returns a list containing the following components:
\item{eigenvalues}{  array_like; \cr
          vector with the eigenvalues.
          }
\item{explvar}{  array_like; \cr
          an vector containing the proportion of variance explained by the first 1, 2,.,k principal components obtained.
          }
\item{loadings}{  array_like; \cr
          the loadings of the principal components.
          }
\item{coord_ind}{  array_like; \cr
          matrix with the coordinates of individuals.
          }
\item{coord_var}{  array_like; \cr
          matrix with the coordinates of variables.
          }
}
\description{
This function performs the representation of GHBiplot (Gabriel,1971).
}
\details{
Algorithm used to construct the GH Biplot. The Biplot is obtained as result of the configuration of markers for individuals and markers for variables in a reference system defined by the factorial axes resulting from the Decomposition in Singular Values (DVS).
}
\examples{
 GHBiplot(mtcars)
}
\references{
\itemize{
 \item Gabriel, K. R. (1971). The Biplot graphic display of matrices with applications to principal components analysis. Biometrika, 58(3), 453-467.
}
}
